+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.1.0] - 2023-07-09
+
+### Added
+
+- `\Szemul\Database\Exception\EntityNotFoundException` Moved from `szemul/framework`
+- `\Szemul\Database\Dao\DaoAbstract` Moved from `szemul/framework`
+
 ## [3.0.2] - 2023-07-02
 
 ### Added